Updated DialogUtils to only allow one AlertDialog to be shown
https://app.asana.com/0/16397576790242/45197032153761

## What Changed
Whenever the CM app fails to load data multiple dialog boxes created
from DialogUtils.quickDialog( ) were shown at once causing an
inconvenience for the user when closing the dialog boxes. Now the
method only allows for one dialog to be created so that this problem
does not happen.

## QA
- [ ] CM:
  - [ ] When all forms of a connection to the internet are turned off,
only one dialog box shows the message that content failed to load.

## Review
- [ ] @brentwatson